+++ core/src/main/java/es/urjc/etsii/grafo/patches/PatchMathRandom.java
@@ -1,16 +1,13 @@
 package es.urjc.etsii.grafo.patches;
 
 import es.urjc.etsii.grafo.config.BlockConfig;
+import jakarta.annotation.PostConstruct;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.objenesis.instantiator.util.UnsafeUtils;
 import org.springframework.stereotype.Service;
 
-import jakarta.annotation.PostConstruct;
-import java.lang.invoke.MethodHandles;
-import java.lang.invoke.VarHandle;
-import java.lang.reflect.Field;
 import java.lang.reflect.InaccessibleObjectException;
-import java.lang.reflect.Modifier;
 
 /**
  * Block calls to Math.random() as it uses an internal random that breaks experiment reproducibility
@@ -43,24 +40,17 @@ public class PatchMathRandom {
         }
 
         try {
-            var internalClass = Class.forName("java.lang.Math$RandomNumberGeneratorHolder");
-            var internalRandom = internalClass.getDeclaredField("randomNumberGenerator");
-            internalRandom.setAccessible(true);
-            makeNonFinal(internalRandom);
-            internalRandom.set(null, new FailRandom());
+            var unsafe = UnsafeUtils.getUnsafe();
+
+            var internalRandomClass = Class.forName("java.lang.Math$RandomNumberGeneratorHolder");
+            var internalRandomField = internalRandomClass.getDeclaredField("randomNumberGenerator");
+            var base = unsafe.staticFieldBase(internalRandomField);
+            var offset = unsafe.staticFieldOffset(internalRandomField);
+            unsafe.putObject(base, offset, new FailRandom());
             log.info("Math.random() patched successfully");
-        } catch (NoSuchFieldException | IllegalAccessException | ClassNotFoundException | InaccessibleObjectException | UnsupportedOperationException e) {
+        } catch (NoSuchFieldException | ClassNotFoundException | InaccessibleObjectException | UnsupportedOperationException e) {
             // Log as warning, but do not stop application when failing to patch, it is not critical
             log.warn("Failed to patch Math.random(), probably due to missing opens, see: https://mork-optimization.readthedocs.io/en/latest/quickstart/troubleshooting/. Cause: {}", e.getMessage());
-        }
-    }
-
-    private void makeNonFinal(Field field) throws NoSuchFieldException, IllegalAccessException {
-        var lookup = MethodHandles.privateLookupIn(Field.class, MethodHandles.lookup());
-        VarHandle modifiers = lookup.findVarHandle(Field.class, "modifiers", int.class);
-        int mods = field.getModifiers();
-        if (Modifier.isFinal(mods)) {
-            modifiers.set(field, mods & ~Modifier.FINAL);
         }
     }
 }